Move timeout at the step level
We recently moved `continue-on-error` from the job, to the step, but
left `timeout-minutes` at the job level; this means that if the steps
error out, the exception is swallowed (as expected), and the job mobes
on; however, if the step / job errors out, that's it, it would cause the
whole workflow to abort.

What we want is to tolerate any kind of errors at the step which we know
might fail, and we try to do that by moving `continue-on-error` and
`timeout-minutes`, together.